Remove all remaining Duplicated code between decksite and logsite.
Fixes #4743
