+++ logsite/__init__.py
@@ -1,13 +1,16 @@
 import subprocess
+from typing import Dict, List, Union
+
+from flask import url_for
+from flask_babel import Babel
 from sqlalchemy import create_engine
 from sqlalchemy_utils import create_database, database_exists
 
-from flask import Flask
-from flask_babel import Babel
 
 from shared import configuration
+from shared_web.flask_app import PDFlask
 
-APP = Flask(__name__)
+APP = PDFlask(__name__)
 BABEL = Babel(APP)
 
 
@@ -23,5 +26,17 @@ def __create_schema() -> None:
 APP.config['commit-id'] = subprocess.check_output(['git', 'rev-parse', 'HEAD'])
 APP.config['branch'] = subprocess.check_output(['git', 'rev-parse', '--abbrev-ref', 'HEAD']).strip().decode()
 APP.config['SECRET_KEY'] = configuration.get('oauth2_client_secret')
+APP.config['js_url'] = 'https://pennydreadfulmagic.com/static/js/pd.js'
+APP.config['css_url'] = 'https://pennydreadfulmagic.com/static/css/pd.css'
+def build_menu() -> List[Dict[str, Union[str, Dict[str, str]]]]:
+    menu = [
+        {'name': 'Home', 'url': url_for('home')},
+        {'name': 'Matches', 'url': url_for('matches')},
+        {'name': 'People', 'url': url_for('people')},
+        {'name': 'About', 'url': url_for('about')},
+    ]
+    return menu
+
+APP.config['menu'] = build_menu
 
 __create_schema()
