+++ shared_web/templates/error.mustache
@@ -2,7 +2,7 @@
 <section>
     <p>{{message}}</p>
     <p><a class="external" href="https://github.com/PennyDreadfulMTG/Penny-Dreadful-Tools/issues">Report bug</a></p>
-    <a href="{{card.url}}" class="card"><img src="{{card.img_url}}" alt="{{card.name}}" class="card-img {{card_img_class}}"></a>
+    <a href="{{card.url}}" class="card"><img src="{{card.img_url}}" alt="{{card.name}}" class="card-img {{c.card_img_class}}"></a>
 </section>
 <section>
     <p>Try these?</p>
