+++ shared_web/views/__init__.py
@@ -1,0 +1,3 @@
+from .internal_server_error import InternalServerError
+from .not_found import NotFound
+from .unauthorized import Unauthorized
